Nio|Epoll|KqueueEventLoop task execution might throw UnsupportedOperationException on shutdown
Motivation:

here is a racy UnsupportedOperationException instead because the task removal is delegated to MpscChunkedArrayQueue that does not support removal. This happens with SingleThreadEventExecutor that overrides the newTaskQueue to return an MPSC queue instead of the LinkedBlockingQueue returned by the base class such as NioEventLoop, EpollEventLoop and KQueueEventLoop.

Modifications:

- Override removeTask(...) where its not supported
- adjust execute(...) to be less racy
- add unit test.

Result:

Fix https://github.com/netty/netty/issues/8475